Bgk 53

+++ BGK.Business/Models/CompanyModel.cs
@@ -6,7 +6,7 @@ namespace BearGoodbyeKolkhozProject.Business.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Tin { get; set; }
+        public long Tin { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public string Password { get; set; }
