+++ BearGoodbyeKolkhoz.API/Models/InputModels/CompanyInputModel.cs
@@ -3,7 +3,7 @@
     public class CompanyInputModel
     {
         public string Name { get; set; }
-        public int Tin { get; set; }
+        public long Tin { get; set; }
         public string Email { get; set; }
         public string PhoneNumber { get; set; }
         public string Password { get; set; }
